Improves subscriber error handling
This commit improves error handling on data subscriber, focusing
on fixing two errors.

The first error condition is related to subscriber attempting to
initialize before the rest of dojot (most notably auth and data-broker)
is up. That would lead the subscriber to a state where it is not
properly initialized, effectively monitoring nothing.

The second is related to errors being raised when handling events
received from an iotagent. On such occasions the process handling
a tenant's message flow would get stuck and no further processing
would occur.